Module arrow_flight::flight_service_server

Low Level tonic FlightServiceServer and FlightService.

Structs

FlightServiceServer

A flight service is an endpoint for retrieving or storing Arrow data. A flight service can expose one or more predefined endpoints that can be accessed using the Arrow Flight Protocol. Additionally, a flight service can expose a set of actions that are available.

Traits

FlightService

Generated trait containing gRPC methods that should be implemented for use with FlightServiceServer.