Struct arrow_flight::gen::FlightData

pub struct FlightData {
    pub flight_descriptor: Option<FlightDescriptor>,
    pub data_header: Bytes,
    pub app_metadata: Bytes,
    pub data_body: Bytes,
}

A batch of Arrow data as part of a stream of batches.

Fields
flight_descriptor: Option<FlightDescriptor>

The descriptor of the data. This is only relevant when a client is starting a new DoPut stream.

data_header: Bytes

Header for message data as described in Message.fbs::Message.

app_metadata: Bytes

Application-defined metadata.

data_body: Bytes

The actual batch of Arrow data. Preferably handled with minimal-copies coming last in the definition to help with sidecar patterns (it is expected that some implementations will fetch this field off the wire with specialized code to avoid extra memory copies).

Implementations

impl FlightData

pub fn new() -> Self

Create a new FlightData.

See Also

See FlightDataEncoderBuilder for a higher level API to convert a stream of RecordBatches to FlightDatas

Example:

// Get encoded Arrow IPC data:
let data_body: Bytes = encode_data();
// Create the FlightData message
let flight_data = FlightData::new()
  .with_descriptor(FlightDescriptor::new_cmd("the command"))
  .with_app_metadata("My apps metadata")
  .with_data_body(data_body);

pub fn with_descriptor(self, flight_descriptor: FlightDescriptor) -> Self

Add a FlightDescriptor describing the data

pub fn with_data_header(self, data_header: impl Into<Bytes>) -> Self

Add a data header

pub fn with_data_body(self, data_body: impl Into<Bytes>) -> Self

Add a data body. See IpcDataGenerator to create this data.

pub fn with_app_metadata(self, app_metadata: impl Into<Bytes>) -> Self

Add optional application specific metadata to the message