arrow_flight::gen

Struct FlightInfo

pub struct FlightInfo {
    pub schema: Bytes,
    pub flight_descriptor: Option<FlightDescriptor>,
    pub endpoint: Vec<FlightEndpoint>,
    pub total_records: i64,
    pub total_bytes: i64,
    pub ordered: bool,
    pub app_metadata: Bytes,
}

The access coordinates for retrieval of a dataset. With a FlightInfo, a consumer is able to determine how to retrieve a dataset.

Fields
schema: Bytes

The schema of the dataset in its IPC form: 4 bytes - an optional IPC_CONTINUATION_TOKEN prefix 4 bytes - the byte length of the payload a flatbuffer Message whose header is the Schema

flight_descriptor: Option<FlightDescriptor>

The descriptor associated with this info.

endpoint: Vec<FlightEndpoint>

A list of endpoints associated with the flight. To consume the whole flight, all endpoints (and hence all Tickets) must be consumed. Endpoints can be consumed in any order.

In other words, an application can use multiple endpoints to represent partitioned data.

If the returned data has an ordering, an application can use โFlightInfo.ordered = trueโ or should return the all data in a single endpoint. Otherwise, there is no ordering defined on endpoints or the data within.

A client can read ordered data by reading data from returned endpoints, in order, from front to back.

Note that a client may ignore โFlightInfo.ordered = trueโ. If an ordering is important for an application, an application must choose one of them:

An application requires that all clients must read data in returned endpoints order.

An application must return the all data in a single endpoint.

total_records: i64

Set these to -1 if unknown.

total_bytes: i64ordered: bool

FlightEndpoints are in the same order as the data.

app_metadata: Bytes

Application-defined metadata.

There is no inherent or required relationship between this and the app_metadata fields in the FlightEndpoints or resulting FlightData messages. Since this metadata is application-defined, a given application could define there to be a relationship, but there is none required by the spec.

Implementations

impl FlightInfo

pub fn new() -> FlightInfo

Create a new, empty FlightInfo, describing where to fetch flight data

Example:

// Create a new FlightInfo
let flight_info = FlightInfo::new()
  // Encode the Arrow schema
  .try_with_schema(&get_schema())
  .expect("encoding failed")
  .with_endpoint(
     FlightEndpoint::new()
       .with_ticket(Ticket::new("ticket contents")
     )
   )
  .with_descriptor(FlightDescriptor::new_cmd("RUN QUERY"));

pub fn try_decode_schema(self) -> Result<Schema, ArrowError>

Try and convert the data in this FlightInfo into a [Schema]

pub fn try_with_schema(self, schema: &Schema) -> Result<Self, ArrowError>

Specify the schema for the response.

Note this takes the arrow [Schema] (not the IPC schema) and encodes it using the default IPC options.

Returns an error if schema can not be encoded into IPC form.

pub fn with_endpoint(self, endpoint: FlightEndpoint) -> Self

Add specific a endpoint for fetching the data

pub fn with_descriptor(self, flight_descriptor: FlightDescriptor) -> Self

Add a FlightDescriptor describing what this data is

pub fn with_total_records(self, total_records: i64) -> Self

Set the number of records in the result, if known

pub fn with_total_bytes(self, total_bytes: i64) -> Self

Set the number of bytes in the result, if known

pub fn with_ordered(self, ordered: bool) -> Self

Specify if the response is ordered across endpoints

pub fn with_app_metadata(self, app_metadata: impl Into<Bytes>) -> Self

Add optional application specific metadata to the message