arrow_flight::sql

Module client

A FlightSQL Client FlightSqlServiceClient

Structs

FlightSqlServiceClient

A FlightSQLServiceClient is an endpoint for retrieving or storing Arrow data by FlightSQL protocol.

PreparedStatement

A PreparedStatement

Enums

ArrowFlightData

A polymorphic structure to natively represent different types of data contained in FlightData

Functions

arrow_data_from_flight_data

Extract Schema or RecordBatches from the FlightData wire representation

decode_error_to_arrow_error 🔒

flight_error_to_arrow_error 🔒

status_to_arrow_error 🔒