arrow_flight::sql

Enum SqlInfo

#[repr(i32)]
pub enum SqlInfo {

    FlightSqlServerName = 0,
    FlightSqlServerVersion = 1,
    FlightSqlServerArrowVersion = 2,
    FlightSqlServerReadOnly = 3,
    FlightSqlServerSql = 4,
    FlightSqlServerSubstrait = 5,
    FlightSqlServerSubstraitMinVersion = 6,
    FlightSqlServerSubstraitMaxVersion = 7,
    FlightSqlServerTransaction = 8,
    FlightSqlServerCancel = 9,
    FlightSqlServerBulkIngestion = 10,
    FlightSqlServerIngestTransactionsSupported = 11,
    FlightSqlServerStatementTimeout = 100,
    FlightSqlServerTransactionTimeout = 101,
    SqlDdlCatalog = 500,
    SqlDdlSchema = 501,
    SqlDdlTable = 502,
    SqlIdentifierCase = 503,
    SqlIdentifierQuoteChar = 504,
    SqlQuotedIdentifierCase = 505,
    SqlAllTablesAreSelectable = 506,
    SqlNullOrdering = 507,
    SqlKeywords = 508,
    SqlNumericFunctions = 509,
    SqlStringFunctions = 510,
    SqlSystemFunctions = 511,
    SqlDatetimeFunctions = 512,
    SqlSearchStringEscape = 513,
    SqlExtraNameCharacters = 514,
    SqlSupportsColumnAliasing = 515,
    SqlNullPlusNullIsNull = 516,
    SqlSupportsConvert = 517,
    SqlSupportsTableCorrelationNames = 518,
    SqlSupportsDifferentTableCorrelationNames = 519,
    SqlSupportsExpressionsInOrderBy = 520,
    SqlSupportsOrderByUnrelated = 521,
    SqlSupportedGroupBy = 522,
    SqlSupportsLikeEscapeClause = 523,
    SqlSupportsNonNullableColumns = 524,
    SqlSupportedGrammar = 525,
    SqlAnsi92SupportedLevel = 526,
    SqlSupportsIntegrityEnhancementFacility = 527,
    SqlOuterJoinsSupportLevel = 528,
    SqlSchemaTerm = 529,
    SqlProcedureTerm = 530,
    SqlCatalogTerm = 531,
    SqlCatalogAtStart = 532,
    SqlSchemasSupportedActions = 533,
    SqlCatalogsSupportedActions = 534,
    SqlSupportedPositionedCommands = 535,
    SqlSelectForUpdateSupported = 536,
    SqlStoredProceduresSupported = 537,
    SqlSupportedSubqueries = 538,
    SqlCorrelatedSubqueriesSupported = 539,
    SqlSupportedUnions = 540,
    SqlMaxBinaryLiteralLength = 541,
    SqlMaxCharLiteralLength = 542,
    SqlMaxColumnNameLength = 543,
    SqlMaxColumnsInGroupBy = 544,
    SqlMaxColumnsInIndex = 545,
    SqlMaxColumnsInOrderBy = 546,
    SqlMaxColumnsInSelect = 547,
    SqlMaxColumnsInTable = 548,
    SqlMaxConnections = 549,
    SqlMaxCursorNameLength = 550,
    SqlMaxIndexLength = 551,
    SqlDbSchemaNameLength = 552,
    SqlMaxProcedureNameLength = 553,
    SqlMaxCatalogNameLength = 554,
    SqlMaxRowSize = 555,
    SqlMaxRowSizeIncludesBlobs = 556,
    SqlMaxStatementLength = 557,
    SqlMaxStatements = 558,
    SqlMaxTableNameLength = 559,
    SqlMaxTablesInSelect = 560,
    SqlMaxUsernameLength = 561,
    SqlDefaultTransactionIsolation = 562,
    SqlTransactionsSupported = 563,
    SqlSupportedTransactionsIsolationLevels = 564,
    SqlDataDefinitionCausesTransactionCommit = 565,
    SqlDataDefinitionsInTransactionsIgnored = 566,
    SqlSupportedResultSetTypes = 567,
    SqlSupportedConcurrenciesForResultSetUnspecified = 568,
    SqlSupportedConcurrenciesForResultSetForwardOnly = 569,
    SqlSupportedConcurrenciesForResultSetScrollSensitive = 570,
    SqlSupportedConcurrenciesForResultSetScrollInsensitive = 571,
    SqlBatchUpdatesSupported = 572,
    SqlSavepointsSupported = 573,
    SqlNamedParametersSupported = 574,
    SqlLocatorsUpdateCopy = 575,
    SqlStoredFunctionsUsingCallSyntaxSupported = 576,

}

Options for CommandGetSqlInfo.

Variants

FlightSqlServerName = 0

Retrieves a UTF-8 string with the name of the Flight SQL Server.

FlightSqlServerVersion = 1

Retrieves a UTF-8 string with the native version of the Flight SQL Server.

FlightSqlServerArrowVersion = 2

Retrieves a UTF-8 string with the Arrow format version of the Flight SQL Server.

FlightSqlServerReadOnly = 3

Retrieves a boolean value indicating whether the Flight SQL Server is read only.

Returns:

false: if read-write

true: if read only

FlightSqlServerSql = 4

Retrieves a boolean value indicating whether the Flight SQL Server supports executing SQL queries.

Note that the absence of this info (as opposed to a false value) does not necessarily mean that SQL is not supported, as this property was not originally defined.

FlightSqlServerSubstrait = 5

Retrieves a boolean value indicating whether the Flight SQL Server supports executing Substrait plans.

FlightSqlServerSubstraitMinVersion = 6

Retrieves a string value indicating the minimum supported Substrait version, or null if Substrait is not supported.

FlightSqlServerSubstraitMaxVersion = 7

Retrieves a string value indicating the maximum supported Substrait version, or null if Substrait is not supported.

FlightSqlServerTransaction = 8

Retrieves an int32 indicating whether the Flight SQL Server supports the BeginTransaction/EndTransaction/BeginSavepoint/EndSavepoint actions.

Even if this is not supported, the database may still support explicit “BEGIN TRANSACTION”/“COMMIT” SQL statements (see SQL_TRANSACTIONS_SUPPORTED); this property is only about whether the server implements the Flight SQL API endpoints.

The possible values are listed in SqlSupportedTransaction.

FlightSqlServerCancel = 9

Retrieves a boolean value indicating whether the Flight SQL Server supports explicit query cancellation (the CancelQuery action).

FlightSqlServerBulkIngestion = 10

Retrieves a boolean value indicating whether the Flight SQL Server supports executing bulk ingestion.

FlightSqlServerIngestTransactionsSupported = 11

Retrieves a boolean value indicating whether transactions are supported for bulk ingestion. If not, invoking the method commit in the context of a bulk ingestion is a noop, and the isolation level is arrow.flight.protocol.sql.SqlTransactionIsolationLevel.TRANSACTION_NONE.

Returns:

false: if bulk ingestion transactions are unsupported;

true: if bulk ingestion transactions are supported.

FlightSqlServerStatementTimeout = 100

Retrieves an int32 indicating the timeout (in milliseconds) for prepared statement handles.

If 0, there is no timeout. Servers should reset the timeout when the handle is used in a command.

FlightSqlServerTransactionTimeout = 101

Retrieves an int32 indicating the timeout (in milliseconds) for transactions, since transactions are not tied to a connection.

If 0, there is no timeout. Servers should reset the timeout when the handle is used in a command.

SqlDdlCatalog = 500

Retrieves a boolean value indicating whether the Flight SQL Server supports CREATE and DROP of catalogs.

Returns:

false: if it doesn’t support CREATE and DROP of catalogs.

true: if it supports CREATE and DROP of catalogs.

SqlDdlSchema = 501

Retrieves a boolean value indicating whether the Flight SQL Server supports CREATE and DROP of schemas.

Returns:

false: if it doesn’t support CREATE and DROP of schemas.

true: if it supports CREATE and DROP of schemas.

SqlDdlTable = 502

Indicates whether the Flight SQL Server supports CREATE and DROP of tables.

Returns:

false: if it doesn’t support CREATE and DROP of tables.

true: if it supports CREATE and DROP of tables.

SqlIdentifierCase = 503

Retrieves a int32 ordinal representing the case sensitivity of catalog, table, schema and table names.

The possible values are listed in arrow.flight.protocol.sql.SqlSupportedCaseSensitivity.

SqlIdentifierQuoteChar = 504

Retrieves a UTF-8 string with the supported character(s) used to surround a delimited identifier.

SqlQuotedIdentifierCase = 505

Retrieves a int32 describing the case sensitivity of quoted identifiers.

The possible values are listed in arrow.flight.protocol.sql.SqlSupportedCaseSensitivity.

SqlAllTablesAreSelectable = 506

Retrieves a boolean value indicating whether all tables are selectable.

Returns:

false: if not all tables are selectable or if none are;

true: if all tables are selectable.

SqlNullOrdering = 507

Retrieves the null ordering.

Returns a int32 ordinal for the null ordering being used, as described in arrow.flight.protocol.sql.SqlNullOrdering.

SqlKeywords = 508

Retrieves a UTF-8 string list with values of the supported keywords.

SqlNumericFunctions = 509

Retrieves a UTF-8 string list with values of the supported numeric functions.

SqlStringFunctions = 510

Retrieves a UTF-8 string list with values of the supported string functions.

SqlSystemFunctions = 511

Retrieves a UTF-8 string list with values of the supported system functions.

SqlDatetimeFunctions = 512

Retrieves a UTF-8 string list with values of the supported datetime functions.

SqlSearchStringEscape = 513

Retrieves the UTF-8 string that can be used to escape wildcard characters. This is the string that can be used to escape ‘’ or ‘%’ in the catalog search parameters that are a pattern (and therefore use one of the wildcard characters). The ‘’ character represents any single character; the ‘%’ character represents any sequence of zero or more characters.

SqlExtraNameCharacters = 514

Retrieves a UTF-8 string with all the “extra” characters that can be used in unquoted identifier names (those beyond a-z, A-Z, 0-9 and _).

SqlSupportsColumnAliasing = 515

Retrieves a boolean value indicating whether column aliasing is supported. If so, the SQL AS clause can be used to provide names for computed columns or to provide alias names for columns as required.

Returns:

false: if column aliasing is unsupported;

true: if column aliasing is supported.

SqlNullPlusNullIsNull = 516

Retrieves a boolean value indicating whether concatenations between null and non-null values being null are supported.

Returns:

false: if concatenations between null and non-null values being null are unsupported;

true: if concatenations between null and non-null values being null are supported.

SqlSupportsConvert = 517

Retrieves a map where the key is the type to convert from and the value is a list with the types to convert to, indicating the supported conversions. Each key and each item on the list value is a value to a predefined type on SqlSupportsConvert enum. The returned map will be: map<int32, list>

SqlSupportsTableCorrelationNames = 518

Retrieves a boolean value indicating whether, when table correlation names are supported, they are restricted to being different from the names of the tables.

Returns:

false: if table correlation names are unsupported;

true: if table correlation names are supported.

SqlSupportsDifferentTableCorrelationNames = 519

Retrieves a boolean value indicating whether, when table correlation names are supported, they are restricted to being different from the names of the tables.

Returns:

false: if different table correlation names are unsupported;

true: if different table correlation names are supported

SqlSupportsExpressionsInOrderBy = 520

Retrieves a boolean value indicating whether expressions in ORDER BY lists are supported.

Returns:

false: if expressions in ORDER BY are unsupported;

true: if expressions in ORDER BY are supported;

SqlSupportsOrderByUnrelated = 521

Retrieves a boolean value indicating whether using a column that is not in the SELECT statement in a GROUP BY clause is supported.

Returns:

false: if using a column that is not in the SELECT statement in a GROUP BY clause is unsupported;

true: if using a column that is not in the SELECT statement in a GROUP BY clause is supported.

SqlSupportedGroupBy = 522

Retrieves the supported GROUP BY commands;

Returns an int32 bitmask value representing the supported commands. The returned bitmask should be parsed in order to retrieve the supported commands.

For instance:

return 0 (\b0) => [] (GROUP BY is unsupported);

return 1 (\b1) => [SQL_GROUP_BY_UNRELATED];

return 2 (\b10) => [SQL_GROUP_BY_BEYOND_SELECT];

return 3 (\b11) => [SQL_GROUP_BY_UNRELATED, SQL_GROUP_BY_BEYOND_SELECT]. Valid GROUP BY types are described under arrow.flight.protocol.sql.SqlSupportedGroupBy.

SqlSupportsLikeEscapeClause = 523

Retrieves a boolean value indicating whether specifying a LIKE escape clause is supported.

Returns:

false: if specifying a LIKE escape clause is unsupported;

true: if specifying a LIKE escape clause is supported.

SqlSupportsNonNullableColumns = 524

Retrieves a boolean value indicating whether columns may be defined as non-nullable.

Returns:

false: if columns cannot be defined as non-nullable;

true: if columns may be defined as non-nullable.

SqlSupportedGrammar = 525

Retrieves the supported SQL grammar level as per the ODBC specification.

Returns an int32 bitmask value representing the supported SQL grammar level. The returned bitmask should be parsed in order to retrieve the supported grammar levels.

For instance:

return 0 (\b0) => [] (SQL grammar is unsupported);

return 1 (\b1) => [SQL_MINIMUM_GRAMMAR];

return 2 (\b10) => [SQL_CORE_GRAMMAR];

return 3 (\b11) => [SQL_MINIMUM_GRAMMAR, SQL_CORE_GRAMMAR];

return 4 (\b100) => [SQL_EXTENDED_GRAMMAR];

return 5 (\b101) => [SQL_MINIMUM_GRAMMAR, SQL_EXTENDED_GRAMMAR];

return 6 (\b110) => [SQL_CORE_GRAMMAR, SQL_EXTENDED_GRAMMAR];

return 7 (\b111) => [SQL_MINIMUM_GRAMMAR, SQL_CORE_GRAMMAR, SQL_EXTENDED_GRAMMAR]. Valid SQL grammar levels are described under arrow.flight.protocol.sql.SupportedSqlGrammar.

SqlAnsi92SupportedLevel = 526

Retrieves the supported ANSI92 SQL grammar level.

Returns an int32 bitmask value representing the supported ANSI92 SQL grammar level. The returned bitmask should be parsed in order to retrieve the supported commands.

For instance:

return 0 (\b0) => [] (ANSI92 SQL grammar is unsupported);

return 1 (\b1) => [ANSI92_ENTRY_SQL];

return 2 (\b10) => [ANSI92_INTERMEDIATE_SQL];

return 3 (\b11) => [ANSI92_ENTRY_SQL, ANSI92_INTERMEDIATE_SQL];

return 4 (\b100) => [ANSI92_FULL_SQL];

return 5 (\b101) => [ANSI92_ENTRY_SQL, ANSI92_FULL_SQL];

return 6 (\b110) => [ANSI92_INTERMEDIATE_SQL, ANSI92_FULL_SQL];

return 7 (\b111) => [ANSI92_ENTRY_SQL, ANSI92_INTERMEDIATE_SQL, ANSI92_FULL_SQL]. Valid ANSI92 SQL grammar levels are described under arrow.flight.protocol.sql.SupportedAnsi92SqlGrammarLevel.

SqlSupportsIntegrityEnhancementFacility = 527

Retrieves a boolean value indicating whether the SQL Integrity Enhancement Facility is supported.

Returns:

false: if the SQL Integrity Enhancement Facility is supported;

true: if the SQL Integrity Enhancement Facility is supported.

SqlOuterJoinsSupportLevel = 528

Retrieves the support level for SQL OUTER JOINs.

Returns a int32 ordinal for the SQL ordering being used, as described in arrow.flight.protocol.sql.SqlOuterJoinsSupportLevel.

SqlSchemaTerm = 529

Retrieves a UTF-8 string with the preferred term for “schema”.

SqlProcedureTerm = 530

Retrieves a UTF-8 string with the preferred term for “procedure”.

SqlCatalogTerm = 531

Retrieves a UTF-8 string with the preferred term for “catalog”. If a empty string is returned its assumed that the server does NOT supports catalogs.

SqlCatalogAtStart = 532

Retrieves a boolean value indicating whether a catalog appears at the start of a fully qualified table name.

false: if a catalog does not appear at the start of a fully qualified table name;

true: if a catalog appears at the start of a fully qualified table name.

SqlSchemasSupportedActions = 533

Retrieves the supported actions for a SQL schema.

Returns an int32 bitmask value representing the supported actions for a SQL schema. The returned bitmask should be parsed in order to retrieve the supported actions for a SQL schema.

For instance:

return 0 (\b0) => [] (no supported actions for SQL schema);

return 1 (\b1) => [SQL_ELEMENT_IN_PROCEDURE_CALLS];

return 2 (\b10) => [SQL_ELEMENT_IN_INDEX_DEFINITIONS];

return 3 (\b11) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_INDEX_DEFINITIONS];

return 4 (\b100) => [SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 5 (\b101) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 6 (\b110) => [SQL_ELEMENT_IN_INDEX_DEFINITIONS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 7 (\b111) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_INDEX_DEFINITIONS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS]. Valid actions for a SQL schema described under arrow.flight.protocol.sql.SqlSupportedElementActions.

SqlCatalogsSupportedActions = 534

Retrieves the supported actions for a SQL schema.

Returns an int32 bitmask value representing the supported actions for a SQL catalog. The returned bitmask should be parsed in order to retrieve the supported actions for a SQL catalog.

For instance:

return 0 (\b0) => [] (no supported actions for SQL catalog);

return 1 (\b1) => [SQL_ELEMENT_IN_PROCEDURE_CALLS];

return 2 (\b10) => [SQL_ELEMENT_IN_INDEX_DEFINITIONS];

return 3 (\b11) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_INDEX_DEFINITIONS];

return 4 (\b100) => [SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 5 (\b101) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 6 (\b110) => [SQL_ELEMENT_IN_INDEX_DEFINITIONS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS];

return 7 (\b111) => [SQL_ELEMENT_IN_PROCEDURE_CALLS, SQL_ELEMENT_IN_INDEX_DEFINITIONS, SQL_ELEMENT_IN_PRIVILEGE_DEFINITIONS]. Valid actions for a SQL catalog are described under arrow.flight.protocol.sql.SqlSupportedElementActions.

SqlSupportedPositionedCommands = 535

Retrieves the supported SQL positioned commands.

Returns an int32 bitmask value representing the supported SQL positioned commands. The returned bitmask should be parsed in order to retrieve the supported SQL positioned commands.

For instance:

return 0 (\b0) => [] (no supported SQL positioned commands);

return 1 (\b1) => [SQL_POSITIONED_DELETE];

return 2 (\b10) => [SQL_POSITIONED_UPDATE];

return 3 (\b11) => [SQL_POSITIONED_DELETE, SQL_POSITIONED_UPDATE]. Valid SQL positioned commands are described under arrow.flight.protocol.sql.SqlSupportedPositionedCommands.

SqlSelectForUpdateSupported = 536

Retrieves a boolean value indicating whether SELECT FOR UPDATE statements are supported.

Returns:

false: if SELECT FOR UPDATE statements are unsupported;

true: if SELECT FOR UPDATE statements are supported.

SqlStoredProceduresSupported = 537

Retrieves a boolean value indicating whether stored procedure calls that use the stored procedure escape syntax are supported.

Returns:

false: if stored procedure calls that use the stored procedure escape syntax are unsupported;

true: if stored procedure calls that use the stored procedure escape syntax are supported.

SqlSupportedSubqueries = 538

Retrieves the supported SQL subqueries.

Returns an int32 bitmask value representing the supported SQL subqueries. The returned bitmask should be parsed in order to retrieve the supported SQL subqueries.

For instance:

return 0 (\b0) => [] (no supported SQL subqueries);

return 1 (\b1) => [SQL_SUBQUERIES_IN_COMPARISONS];

return 2 (\b10) => [SQL_SUBQUERIES_IN_EXISTS];

return 3 (\b11) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_EXISTS];

return 4 (\b100) => [SQL_SUBQUERIES_IN_INS];

return 5 (\b101) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_INS];

return 6 (\b110) => [SQL_SUBQUERIES_IN_INS, SQL_SUBQUERIES_IN_EXISTS];

return 7 (\b111) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_EXISTS, SQL_SUBQUERIES_IN_INS];

return 8 (\b1000) => [SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 9 (\b1001) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 10 (\b1010) => [SQL_SUBQUERIES_IN_EXISTS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 11 (\b1011) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_EXISTS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 12 (\b1100) => [SQL_SUBQUERIES_IN_INS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 13 (\b1101) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_INS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 14 (\b1110) => [SQL_SUBQUERIES_IN_EXISTS, SQL_SUBQUERIES_IN_INS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

return 15 (\b1111) => [SQL_SUBQUERIES_IN_COMPARISONS, SQL_SUBQUERIES_IN_EXISTS, SQL_SUBQUERIES_IN_INS, SQL_SUBQUERIES_IN_QUANTIFIEDS];

… Valid SQL subqueries are described under arrow.flight.protocol.sql.SqlSupportedSubqueries.

SqlCorrelatedSubqueriesSupported = 539

Retrieves a boolean value indicating whether correlated subqueries are supported.

Returns:

false: if correlated subqueries are unsupported;

true: if correlated subqueries are supported.

SqlSupportedUnions = 540

Retrieves the supported SQL UNIONs.

Returns an int32 bitmask value representing the supported SQL UNIONs. The returned bitmask should be parsed in order to retrieve the supported SQL UNIONs.

For instance:

return 0 (\b0) => [] (no supported SQL positioned commands);

return 1 (\b1) => [SQL_UNION];

return 2 (\b10) => [SQL_UNION_ALL];

return 3 (\b11) => [SQL_UNION, SQL_UNION_ALL]. Valid SQL positioned commands are described under arrow.flight.protocol.sql.SqlSupportedUnions.

SqlMaxBinaryLiteralLength = 541

Retrieves a int64 value representing the maximum number of hex characters allowed in an inline binary literal.

SqlMaxCharLiteralLength = 542

Retrieves a int64 value representing the maximum number of characters allowed for a character literal.

SqlMaxColumnNameLength = 543

Retrieves a int64 value representing the maximum number of characters allowed for a column name.

SqlMaxColumnsInGroupBy = 544

Retrieves a int64 value representing the maximum number of columns allowed in a GROUP BY clause.

SqlMaxColumnsInIndex = 545

Retrieves a int64 value representing the maximum number of columns allowed in an index.

SqlMaxColumnsInOrderBy = 546

Retrieves a int64 value representing the maximum number of columns allowed in an ORDER BY clause.

SqlMaxColumnsInSelect = 547

Retrieves a int64 value representing the maximum number of columns allowed in a SELECT list.

SqlMaxColumnsInTable = 548

Retrieves a int64 value representing the maximum number of columns allowed in a table.

SqlMaxConnections = 549

Retrieves a int64 value representing the maximum number of concurrent connections possible.

SqlMaxCursorNameLength = 550

Retrieves a int64 value the maximum number of characters allowed in a cursor name.

SqlMaxIndexLength = 551

Retrieves a int64 value representing the maximum number of bytes allowed for an index, including all of the parts of the index.

SqlDbSchemaNameLength = 552

Retrieves a int64 value representing the maximum number of characters allowed in a schema name.

SqlMaxProcedureNameLength = 553

Retrieves a int64 value representing the maximum number of characters allowed in a procedure name.

SqlMaxCatalogNameLength = 554

Retrieves a int64 value representing the maximum number of characters allowed in a catalog name.

SqlMaxRowSize = 555

Retrieves a int64 value representing the maximum number of bytes allowed in a single row.

SqlMaxRowSizeIncludesBlobs = 556

Retrieves a boolean indicating whether the return value for the JDBC method getMaxRowSize includes the SQL data types LONGVARCHAR and LONGVARBINARY.

Returns:

false: if return value for the JDBC method getMaxRowSize does not include the SQL data types LONGVARCHAR and LONGVARBINARY;

true: if return value for the JDBC method getMaxRowSize includes the SQL data types LONGVARCHAR and LONGVARBINARY.

SqlMaxStatementLength = 557

Retrieves a int64 value representing the maximum number of characters allowed for an SQL statement; a result of 0 (zero) means that there is no limit or the limit is not known.

SqlMaxStatements = 558

Retrieves a int64 value representing the maximum number of active statements that can be open at the same time.

SqlMaxTableNameLength = 559

Retrieves a int64 value representing the maximum number of characters allowed in a table name.

SqlMaxTablesInSelect = 560

Retrieves a int64 value representing the maximum number of tables allowed in a SELECT statement.

SqlMaxUsernameLength = 561

Retrieves a int64 value representing the maximum number of characters allowed in a user name.

SqlDefaultTransactionIsolation = 562

Retrieves this database’s default transaction isolation level as described in arrow.flight.protocol.sql.SqlTransactionIsolationLevel.

Returns a int32 ordinal for the SQL transaction isolation level.

SqlTransactionsSupported = 563

Retrieves a boolean value indicating whether transactions are supported. If not, invoking the method commit is a noop, and the isolation level is arrow.flight.protocol.sql.SqlTransactionIsolationLevel.TRANSACTION_NONE.

Returns:

false: if transactions are unsupported;

true: if transactions are supported.

SqlSupportedTransactionsIsolationLevels = 564

Retrieves the supported transactions isolation levels.

Returns an int32 bitmask value representing the supported transactions isolation levels. The returned bitmask should be parsed in order to retrieve the supported transactions isolation levels.

For instance:

return 0 (\b0) => [] (no supported SQL transactions isolation levels);

return 1 (\b1) => [SQL_TRANSACTION_NONE];

return 2 (\b10) => [SQL_TRANSACTION_READ_UNCOMMITTED];

return 3 (\b11) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_READ_UNCOMMITTED];

return 4 (\b100) => [SQL_TRANSACTION_REPEATABLE_READ];

return 5 (\b101) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_REPEATABLE_READ];

return 6 (\b110) => [SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ];

return 7 (\b111) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ];

return 8 (\b1000) => [SQL_TRANSACTION_REPEATABLE_READ];

return 9 (\b1001) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_REPEATABLE_READ];

return 10 (\b1010) => [SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ];

return 11 (\b1011) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ];

return 12 (\b1100) => [SQL_TRANSACTION_REPEATABLE_READ, SQL_TRANSACTION_REPEATABLE_READ];

return 13 (\b1101) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_REPEATABLE_READ, SQL_TRANSACTION_REPEATABLE_READ];

return 14 (\b1110) => [SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ, SQL_TRANSACTION_REPEATABLE_READ];

return 15 (\b1111) => [SQL_TRANSACTION_NONE, SQL_TRANSACTION_READ_UNCOMMITTED, SQL_TRANSACTION_REPEATABLE_READ, SQL_TRANSACTION_REPEATABLE_READ];

return 16 (\b10000) => [SQL_TRANSACTION_SERIALIZABLE];

… Valid SQL positioned commands are described under arrow.flight.protocol.sql.SqlTransactionIsolationLevel.

SqlDataDefinitionCausesTransactionCommit = 565

Retrieves a boolean value indicating whether a data definition statement within a transaction forces the transaction to commit.

Returns:

false: if a data definition statement within a transaction does not force the transaction to commit;

true: if a data definition statement within a transaction forces the transaction to commit.

SqlDataDefinitionsInTransactionsIgnored = 566

Retrieves a boolean value indicating whether a data definition statement within a transaction is ignored.

Returns:

false: if a data definition statement within a transaction is taken into account;

true: a data definition statement within a transaction is ignored.

SqlSupportedResultSetTypes = 567

Retrieves an int32 bitmask value representing the supported result set types. The returned bitmask should be parsed in order to retrieve the supported result set types.

For instance:

return 0 (\b0) => [] (no supported result set types);

return 1 (\b1) => [SQL_RESULT_SET_TYPE_UNSPECIFIED];

return 2 (\b10) => [SQL_RESULT_SET_TYPE_FORWARD_ONLY];

return 3 (\b11) => [SQL_RESULT_SET_TYPE_UNSPECIFIED, SQL_RESULT_SET_TYPE_FORWARD_ONLY];

return 4 (\b100) => [SQL_RESULT_SET_TYPE_SCROLL_INSENSITIVE];

return 5 (\b101) => [SQL_RESULT_SET_TYPE_UNSPECIFIED, SQL_RESULT_SET_TYPE_SCROLL_INSENSITIVE];

return 6 (\b110) => [SQL_RESULT_SET_TYPE_FORWARD_ONLY, SQL_RESULT_SET_TYPE_SCROLL_INSENSITIVE];

return 7 (\b111) => [SQL_RESULT_SET_TYPE_UNSPECIFIED, SQL_RESULT_SET_TYPE_FORWARD_ONLY, SQL_RESULT_SET_TYPE_SCROLL_INSENSITIVE];

return 8 (\b1000) => [SQL_RESULT_SET_TYPE_SCROLL_SENSITIVE];

… Valid result set types are described under arrow.flight.protocol.sql.SqlSupportedResultSetType.

SqlSupportedConcurrenciesForResultSetUnspecified = 568

Returns an int32 bitmask value concurrency types supported for arrow.flight.protocol.sql.SqlSupportedResultSetType.SQL_RESULT_SET_TYPE_UNSPECIFIED.

For instance:

return 0 (\b0) => [] (no supported concurrency types for this result set type)

return 1 (\b1) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED]

return 2 (\b10) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 3 (\b11) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 4 (\b100) => [SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 5 (\b101) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 6 (\b110) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 7 (\b111) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE] Valid result set types are described under arrow.flight.protocol.sql.SqlSupportedResultSetConcurrency.

SqlSupportedConcurrenciesForResultSetForwardOnly = 569

Returns an int32 bitmask value concurrency types supported for arrow.flight.protocol.sql.SqlSupportedResultSetType.SQL_RESULT_SET_TYPE_FORWARD_ONLY.

For instance:

return 0 (\b0) => [] (no supported concurrency types for this result set type)

return 1 (\b1) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED]

return 2 (\b10) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 3 (\b11) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 4 (\b100) => [SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 5 (\b101) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 6 (\b110) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 7 (\b111) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE] Valid result set types are described under arrow.flight.protocol.sql.SqlSupportedResultSetConcurrency.

SqlSupportedConcurrenciesForResultSetScrollSensitive = 570

Returns an int32 bitmask value concurrency types supported for arrow.flight.protocol.sql.SqlSupportedResultSetType.SQL_RESULT_SET_TYPE_SCROLL_SENSITIVE.

For instance:

return 0 (\b0) => [] (no supported concurrency types for this result set type)

return 1 (\b1) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED]

return 2 (\b10) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 3 (\b11) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 4 (\b100) => [SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 5 (\b101) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 6 (\b110) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 7 (\b111) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE] Valid result set types are described under arrow.flight.protocol.sql.SqlSupportedResultSetConcurrency.

SqlSupportedConcurrenciesForResultSetScrollInsensitive = 571

Returns an int32 bitmask value concurrency types supported for arrow.flight.protocol.sql.SqlSupportedResultSetType.SQL_RESULT_SET_TYPE_SCROLL_INSENSITIVE.

For instance:

return 0 (\b0) => [] (no supported concurrency types for this result set type)

return 1 (\b1) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED]

return 2 (\b10) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 3 (\b11) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY]

return 4 (\b100) => [SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 5 (\b101) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 6 (\b110) => [SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE]

return 7 (\b111) => [SQL_RESULT_SET_CONCURRENCY_UNSPECIFIED, SQL_RESULT_SET_CONCURRENCY_READ_ONLY, SQL_RESULT_SET_CONCURRENCY_UPDATABLE] Valid result set types are described under arrow.flight.protocol.sql.SqlSupportedResultSetConcurrency.

SqlBatchUpdatesSupported = 572

Retrieves a boolean value indicating whether this database supports batch updates.

false: if this database does not support batch updates;

true: if this database supports batch updates.

SqlSavepointsSupported = 573

Retrieves a boolean value indicating whether this database supports savepoints.

Returns:

false: if this database does not support savepoints;

true: if this database supports savepoints.

SqlNamedParametersSupported = 574

Retrieves a boolean value indicating whether named parameters are supported in callable statements.

Returns:

false: if named parameters in callable statements are unsupported;

true: if named parameters in callable statements are supported.

SqlLocatorsUpdateCopy = 575

Retrieves a boolean value indicating whether updates made to a LOB are made on a copy or directly to the LOB.

Returns:

false: if updates made to a LOB are made directly to the LOB;

true: if updates made to a LOB are made on a copy.

SqlStoredFunctionsUsingCallSyntaxSupported = 576

Retrieves a boolean value indicating whether invoking user-defined or vendor functions using the stored procedure escape syntax is supported.

Returns:

false: if invoking user-defined or vendor functions using the stored procedure escape syntax is unsupported;

true: if invoking user-defined or vendor functions using the stored procedure escape syntax is supported.

Implementations

impl SqlInfo

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of SqlInfo.

pub fn from_i32(value: i32) -> Option<SqlInfo>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a SqlInfo, or None if value is not a valid variant.

impl SqlInfo

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.