Module arrow_flight::sql::gen

Modules

action_cancel_query_result

Nested message and enum types in ActionCancelQueryResult.

action_end_savepoint_request

Nested message and enum types in ActionEndSavepointRequest.

action_end_transaction_request

Nested message and enum types in ActionEndTransactionRequest.

Structs

ActionBeginSavepointRequest

Request message for the “BeginSavepoint” action. Creates a savepoint within a transaction.

ActionBeginSavepointResult

The result of a “BeginSavepoint” action.

ActionBeginTransactionRequest

Request message for the “BeginTransaction” action. Begins a transaction.

ActionBeginTransactionResult

The result of a “BeginTransaction” action.

ActionCancelQueryRequest

Request message for the “CancelQuery” action.

ActionCancelQueryResult

The result of cancelling a query.

ActionClosePreparedStatementRequest

Request message for the “ClosePreparedStatement” action on a Flight SQL enabled backend. Closes server resources associated with the prepared statement handle.

ActionCreatePreparedStatementRequest

Request message for the “CreatePreparedStatement” action on a Flight SQL enabled backend.

ActionCreatePreparedStatementResult

Wrap the result of a “CreatePreparedStatement” or “CreatePreparedSubstraitPlan” action.

ActionCreatePreparedSubstraitPlanRequest

Request message for the “CreatePreparedSubstraitPlan” action on a Flight SQL enabled backend.

ActionEndSavepointRequest

Request message for the “EndSavepoint” action.

ActionEndTransactionRequest

Request message for the “EndTransaction” action.

CommandGetCatalogs

Represents a request to retrieve the list of catalogs on a Flight SQL enabled backend. The definition of a catalog depends on vendor/implementation. It is usually the database itself Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetCrossReference

Represents a request to retrieve a description of the foreign key columns in the given foreign key table that reference the primary key or the columns representing a unique constraint of the parent table (could be the same or a different table) on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetDbSchemas

Represents a request to retrieve the list of database schemas on a Flight SQL enabled backend. The definition of a database schema depends on vendor/implementation. It is usually a collection of tables. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetExportedKeys

Represents a request to retrieve a description of the foreign key columns that reference the given table’s primary key columns (the foreign keys exported by a table) of a table on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetImportedKeys

Represents a request to retrieve the foreign keys of a table on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetPrimaryKeys

Represents a request to retrieve the primary keys of a table on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetSqlInfo

Represents a metadata request. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetTableTypes

Represents a request to retrieve the list of table types on a Flight SQL enabled backend. The table types depend on vendor/implementation. It is usually used to separate tables from views or system tables. TABLE, VIEW, and SYSTEM TABLE are commonly supported. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetTables

Represents a request to retrieve the list of tables, and optionally their schemas, on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandGetXdbcTypeInfo

Represents a request to retrieve information about data type supported on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

CommandPreparedStatementQuery

Represents an instance of executing a prepared statement. Used in the command member of FlightDescriptor for the following RPC calls:

CommandPreparedStatementUpdate

Represents a SQL update query. Used in the command member of FlightDescriptor for the RPC call DoPut to cause the server to execute the included prepared statement handle as an update.

CommandStatementQuery

Represents a SQL query. Used in the command member of FlightDescriptor for the following RPC calls:

CommandStatementSubstraitPlan

Represents a Substrait plan. Used in the command member of FlightDescriptor for the following RPC calls:

CommandStatementUpdate

Represents a SQL update query. Used in the command member of FlightDescriptor for the RPC call DoPut to cause the server to execute the included SQL update.

DoPutPreparedStatementResult

An optional response returned when DoPut is called with CommandPreparedStatementQuery.

DoPutUpdateResult

Returned from the RPC call DoPut when a CommandStatementUpdate CommandPreparedStatementUpdate was in the request, containing results from the update.

SubstraitPlan

An embedded message describing a Substrait plan to execute.

TicketStatementQuery

Enums

Nullable

Searchable

SqlInfo

Options for CommandGetSqlInfo.

SqlNullOrdering

SqlOuterJoinsSupportLevel

SqlSupportedCaseSensitivity

SqlSupportedElementActions

SqlSupportedGroupBy

SqlSupportedPositionedCommands

SqlSupportedResultSetConcurrency

SqlSupportedResultSetType

SqlSupportedSubqueries

SqlSupportedTransaction

The level of support for Flight SQL transaction RPCs.

SqlSupportedTransactions

SqlSupportedUnions

SqlSupportsConvert

SqlTransactionIsolationLevel

SupportedAnsi92SqlGrammarLevel

SupportedSqlGrammar

UpdateDeleteRules

XdbcDataType

XdbcDatetimeSubcode