arrow_flight::sql::metadata

Struct SqlInfoDataBuilder

pub struct SqlInfoDataBuilder {
    infos: BTreeMap<u32, SqlInfoValue>,
}

Helper to create CommandGetSqlInfo responses.

CommandGetSqlInfo are metadata requests used by a Flight SQL server to communicate supported capabilities to Flight SQL clients.

Servers constuct - usually static - SqlInfoData via the SqlInfoDataBuilder, and build responses using CommandGetSqlInfo::into_builder

Fields
infos: BTreeMap<u32, SqlInfoValue>

Use BTreeMap to ensure the values are sorted by value as to make output consistent

Use u32 to support โcustomโ sql info values that are not part of the SqlInfo enum

Implementations

impl SqlInfoDataBuilder

pub fn new() -> Self

Create a new SQL info builder

pub fn append(&mut self, name: impl SqlInfoName, value: impl Into<SqlInfoValue>)

register the specific sql metadata item

pub fn build(self) -> Result<SqlInfoData>

Encode the contents of this list according to the FlightSQL spec

pub fn schema() -> &'static Schema

Return the [Schema] for a GetSchema RPC call with crate::sql::CommandGetSqlInfo