Trait arrow_flight::sql::server::FlightSqlService

pub trait FlightSqlService: Sync + Send + Sized + 'static {
    type FlightService: FlightService;

    // Required methods
    fn get_flight_info_statement<'life0, 'async_trait>(
        &'life0 self,
        query: CommandStatementQuery,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_substrait_plan<'life0, 'async_trait>(
        &'life0 self,
        query: CommandStatementSubstraitPlan,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_prepared_statement<'life0, 'async_trait>(
        &'life0 self,
        query: CommandPreparedStatementQuery,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_catalogs<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetCatalogs,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_schemas<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetDbSchemas,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_tables<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetTables,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_table_types<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetTableTypes,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_sql_info<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetSqlInfo,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_primary_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetPrimaryKeys,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_exported_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetExportedKeys,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_imported_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetImportedKeys,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_cross_reference<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetCrossReference,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flight_info_xdbc_type_info<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetXdbcTypeInfo,
        request: Request<FlightDescriptor>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_statement<'life0, 'async_trait>(
        &'life0 self,
        ticket: TicketStatementQuery,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_prepared_statement<'life0, 'async_trait>(
        &'life0 self,
        query: CommandPreparedStatementQuery,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_catalogs<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetCatalogs,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_schemas<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetDbSchemas,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_tables<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetTables,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_table_types<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetTableTypes,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_sql_info<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetSqlInfo,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_primary_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetPrimaryKeys,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_exported_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetExportedKeys,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_imported_keys<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetImportedKeys,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_cross_reference<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetCrossReference,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_get_xdbc_type_info<'life0, 'async_trait>(
        &'life0 self,
        query: CommandGetXdbcTypeInfo,
        request: Request<Ticket>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_put_statement_update<'life0, 'async_trait>(
        &'life0 self,
        ticket: CommandStatementUpdate,
        request: Request<Streaming<FlightData>>
    ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_put_prepared_statement_query<'life0, 'async_trait>(
        &'life0 self,
        query: CommandPreparedStatementQuery,
        request: Request<Streaming<FlightData>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoPutStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_put_prepared_statement_update<'life0, 'async_trait>(
        &'life0 self,
        query: CommandPreparedStatementUpdate,
        request: Request<Streaming<FlightData>>
    ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_put_substrait_plan<'life0, 'async_trait>(
        &'life0 self,
        query: CommandStatementSubstraitPlan,
        request: Request<Streaming<FlightData>>
    ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_create_prepared_statement<'life0, 'async_trait>(
        &'life0 self,
        query: ActionCreatePreparedStatementRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<ActionCreatePreparedStatementResult, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_close_prepared_statement<'life0, 'async_trait>(
        &'life0 self,
        query: ActionClosePreparedStatementRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_create_prepared_substrait_plan<'life0, 'async_trait>(
        &'life0 self,
        query: ActionCreatePreparedSubstraitPlanRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<ActionCreatePreparedStatementResult, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_begin_transaction<'life0, 'async_trait>(
        &'life0 self,
        query: ActionBeginTransactionRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<ActionBeginTransactionResult, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_end_transaction<'life0, 'async_trait>(
        &'life0 self,
        query: ActionEndTransactionRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_begin_savepoint<'life0, 'async_trait>(
        &'life0 self,
        query: ActionBeginSavepointRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<ActionBeginSavepointResult, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_end_savepoint<'life0, 'async_trait>(
        &'life0 self,
        query: ActionEndSavepointRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn do_action_cancel_query<'life0, 'async_trait>(
        &'life0 self,
        query: ActionCancelQueryRequest,
        request: Request<Action>
    ) -> Pin<Box<dyn Future<Output = Result<ActionCancelQueryResult, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_sql_info<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: i32,
        result: &'life1 SqlInfo
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn do_handshake<'life0, 'async_trait>(
        &'life0 self,
        _request: Request<Streaming<HandshakeRequest>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<HandshakeResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn do_get_fallback<'life0, 'async_trait>(
        &'life0 self,
        _request: Request<Ticket>,
        message: Any
    ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }

}

Implements FlightSqlService to handle the flight sql protocol

Required Associated Types

type FlightService: FlightService

When impl FlightSqlService, you can always set FlightService to Self

Required Methods

fn get_flight_info_statement<'life0, 'async_trait>( &'life0 self, query: CommandStatementQuery, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for executing a SQL query.

fn get_flight_info_substrait_plan<'life0, 'async_trait>( &'life0 self, query: CommandStatementSubstraitPlan, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for executing a substrait plan.

fn get_flight_info_prepared_statement<'life0, 'async_trait>( &'life0 self, query: CommandPreparedStatementQuery, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for executing an already created prepared statement.

fn get_flight_info_catalogs<'life0, 'async_trait>( &'life0 self, query: CommandGetCatalogs, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for listing catalogs.

fn get_flight_info_schemas<'life0, 'async_trait>( &'life0 self, query: CommandGetDbSchemas, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for listing schemas.

fn get_flight_info_tables<'life0, 'async_trait>( &'life0 self, query: CommandGetTables, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for listing tables.

fn get_flight_info_table_types<'life0, 'async_trait>( &'life0 self, query: CommandGetTableTypes, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about the table types.

fn get_flight_info_sql_info<'life0, 'async_trait>( &'life0 self, query: CommandGetSqlInfo, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo for retrieving other information (See SqlInfo).

fn get_flight_info_primary_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetPrimaryKeys, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about primary and foreign keys.

fn get_flight_info_exported_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetExportedKeys, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about exported keys.

fn get_flight_info_imported_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetImportedKeys, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about imported keys.

fn get_flight_info_cross_reference<'life0, 'async_trait>( &'life0 self, query: CommandGetCrossReference, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about cross reference.

fn get_flight_info_xdbc_type_info<'life0, 'async_trait>( &'life0 self, query: CommandGetXdbcTypeInfo, request: Request<FlightDescriptor> ) -> Pin<Box<dyn Future<Output = Result<Response<FlightInfo>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightInfo to extract information about the supported XDBC types.

fn do_get_statement<'life0, 'async_trait>( &'life0 self, ticket: TicketStatementQuery, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the query results.

fn do_get_prepared_statement<'life0, 'async_trait>( &'life0 self, query: CommandPreparedStatementQuery, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the prepared statement query results.

fn do_get_catalogs<'life0, 'async_trait>( &'life0 self, query: CommandGetCatalogs, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the list of catalogs.

fn do_get_schemas<'life0, 'async_trait>( &'life0 self, query: CommandGetDbSchemas, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the list of schemas.

fn do_get_tables<'life0, 'async_trait>( &'life0 self, query: CommandGetTables, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the list of tables.

fn do_get_table_types<'life0, 'async_trait>( &'life0 self, query: CommandGetTableTypes, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the table types.

fn do_get_sql_info<'life0, 'async_trait>( &'life0 self, query: CommandGetSqlInfo, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the list of SqlInfo results.

fn do_get_primary_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetPrimaryKeys, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the primary and foreign keys.

fn do_get_exported_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetExportedKeys, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the exported keys.

fn do_get_imported_keys<'life0, 'async_trait>( &'life0 self, query: CommandGetImportedKeys, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the imported keys.

fn do_get_cross_reference<'life0, 'async_trait>( &'life0 self, query: CommandGetCrossReference, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the cross reference.

fn do_get_xdbc_type_info<'life0, 'async_trait>( &'life0 self, query: CommandGetXdbcTypeInfo, request: Request<Ticket> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get a FlightDataStream containing the data related to the supported XDBC types.

fn do_put_statement_update<'life0, 'async_trait>( &'life0 self, ticket: CommandStatementUpdate, request: Request<Streaming<FlightData>> ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Execute an update SQL statement.

fn do_put_prepared_statement_query<'life0, 'async_trait>( &'life0 self, query: CommandPreparedStatementQuery, request: Request<Streaming<FlightData>> ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoPutStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Bind parameters to given prepared statement.

fn do_put_prepared_statement_update<'life0, 'async_trait>( &'life0 self, query: CommandPreparedStatementUpdate, request: Request<Streaming<FlightData>> ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Execute an update SQL prepared statement.

fn do_put_substrait_plan<'life0, 'async_trait>( &'life0 self, query: CommandStatementSubstraitPlan, request: Request<Streaming<FlightData>> ) -> Pin<Box<dyn Future<Output = Result<i64, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Execute a substrait plan

fn do_action_create_prepared_statement<'life0, 'async_trait>( &'life0 self, query: ActionCreatePreparedStatementRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<ActionCreatePreparedStatementResult, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a prepared statement from given SQL statement.

fn do_action_close_prepared_statement<'life0, 'async_trait>( &'life0 self, query: ActionClosePreparedStatementRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Close a prepared statement.

fn do_action_create_prepared_substrait_plan<'life0, 'async_trait>( &'life0 self, query: ActionCreatePreparedSubstraitPlanRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<ActionCreatePreparedStatementResult, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a prepared substrait plan.

fn do_action_begin_transaction<'life0, 'async_trait>( &'life0 self, query: ActionBeginTransactionRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<ActionBeginTransactionResult, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Begin a transaction

fn do_action_end_transaction<'life0, 'async_trait>( &'life0 self, query: ActionEndTransactionRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

End a transaction

fn do_action_begin_savepoint<'life0, 'async_trait>( &'life0 self, query: ActionBeginSavepointRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<ActionBeginSavepointResult, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Begin a savepoint

fn do_action_end_savepoint<'life0, 'async_trait>( &'life0 self, query: ActionEndSavepointRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<(), Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

End a savepoint

fn do_action_cancel_query<'life0, 'async_trait>( &'life0 self, query: ActionCancelQueryRequest, request: Request<Action> ) -> Pin<Box<dyn Future<Output = Result<ActionCancelQueryResult, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Cancel a query

fn register_sql_info<'life0, 'life1, 'async_trait>( &'life0 self, id: i32, result: &'life1 SqlInfo ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Register a new SqlInfo result, making it available when calling GetSqlInfo.

Provided Methods

fn do_handshake<'life0, 'async_trait>( &'life0 self, _request: Request<Streaming<HandshakeRequest>> ) -> Pin<Box<dyn Future<Output = Result<Response<Pin<Box<dyn Stream<Item = Result<HandshakeResponse, Status>> + Send>>>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Accept authentication and return a token https://arrow.apache.org/docs/format/Flight.html#authentication

fn do_get_fallback<'life0, 'async_trait>( &'life0 self, _request: Request<Ticket>, message: Any ) -> Pin<Box<dyn Future<Output = Result<Response<<Self as FlightService>::DoGetStream>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Implementors may override to handle additional calls to do_get()