arrow_flight::sql

Struct CommandGetTables

pub struct CommandGetTables {
    pub catalog: Option<String>,
    pub db_schema_filter_pattern: Option<String>,
    pub table_name_filter_pattern: Option<String>,
    pub table_types: Vec<String>,
    pub include_schema: bool,
}

Represents a request to retrieve the list of tables, and optionally their schemas, on a Flight SQL enabled backend. Used in the command member of FlightDescriptor for the following RPC calls:

GetSchema: return the Arrow schema of the query.

GetFlightInfo: execute the catalog metadata request.

The returned Arrow schema will be: < catalog_name: utf8, db_schema_name: utf8, table_name: utf8 not null, table_type: utf8 not null, [optional] table_schema: bytes not null (schema of the table as described in Schema.fbs::Schema, it is serialized as an IPC message.)

Fields on table_schema may contain the following metadata:

ARROW:FLIGHT:SQL:CATALOG_NAME - Tableโs catalog name

ARROW:FLIGHT:SQL:DB_SCHEMA_NAME - Database schema name

ARROW:FLIGHT:SQL:TABLE_NAME - Table name

ARROW:FLIGHT:SQL:TYPE_NAME - The data source-specific name for the data type of the column.

ARROW:FLIGHT:SQL:PRECISION - Column precision/size

ARROW:FLIGHT:SQL:SCALE - Column scale/decimal digits if applicable

ARROW:FLIGHT:SQL:IS_AUTO_INCREMENT - โ1โ indicates if the column is auto incremented, โ0โ otherwise.

ARROW:FLIGHT:SQL:IS_CASE_SENSITIVE - โ1โ indicates if the column is case-sensitive, โ0โ otherwise.

ARROW:FLIGHT:SQL:IS_READ_ONLY - โ1โ indicates if the column is read only, โ0โ otherwise.

ARROW:FLIGHT:SQL:IS_SEARCHABLE - โ1โ indicates if the column is searchable via WHERE clause, โ0โ otherwise. The returned data should be ordered by catalog_name, db_schema_name, table_name, then table_type, followed by table_schema if requested.

Fields
catalog: Option<String>

Specifies the Catalog to search for the tables. An empty string retrieves those without a catalog. If omitted the catalog name should not be used to narrow the search.

db_schema_filter_pattern: Option<String>

Specifies a filter pattern for schemas to search for. When no db_schema_filter_pattern is provided, all schemas matching other filters are searched. In the pattern string, two special characters can be used to denote matching rules: - โ%โ means to match any substring with 0 or more characters. - โ_โ means to match any one character.

table_name_filter_pattern: Option<String>

Specifies a filter pattern for tables to search for. When no table_name_filter_pattern is provided, all tables matching other filters are searched. In the pattern string, two special characters can be used to denote matching rules: - โ%โ means to match any substring with 0 or more characters. - โ_โ means to match any one character.

table_types: Vec<String>

Specifies a filter of table types which must match. The table types depend on vendor/implementation. It is usually used to separate tables from views or system tables. TABLE, VIEW, and SYSTEM TABLE are commonly supported.

include_schema: bool

Specifies if the Arrow schema should be returned for found tables.

Implementations

impl CommandGetTables

pub fn catalog(&self) -> &str

Returns the value of catalog, or the default value if catalog is unset.

pub fn db_schema_filter_pattern(&self) -> &str

Returns the value of db_schema_filter_pattern, or the default value if db_schema_filter_pattern is unset.

pub fn table_name_filter_pattern(&self) -> &str

Returns the value of table_name_filter_pattern, or the default value if table_name_filter_pattern is unset.

impl CommandGetTables

pub fn into_builder(self) -> GetTablesBuilder

Create a builder suitable for constructing a response