Module arrow_flight::utils

Utilities to assist with reading and writing Arrow data as Flight messages

Functions

batches_to_flight_data

Convert RecordBatches to wire protocol FlightDatas

flight_data_from_arrow_batchDeprecated

Convert a RecordBatch to a vector of FlightData representing the bytes of the dictionaries and a FlightData representing the bytes of the batch’s values

flight_data_from_arrow_schemaDeprecated

Convert a Schema to FlightData by converting to an IPC message

flight_data_to_arrow_batch

Convert FlightData (with supplied schema and dictionaries) to an arrow RecordBatch.

flight_data_to_batches

Convert a slice of wire protocol FlightDatas into a vector of RecordBatches

flight_schema_from_arrow_schemaDeprecated

Convert a Schema to SchemaResult by converting to an IPC message

ipc_message_from_arrow_schemaDeprecated

Convert a Schema to bytes in the format expected in FlightInfo.schema