Crate arrow_flight

A native Rust implementation of Apache Arrow Flight for exchanging Arrow data between processes.

Please see the arrow-flight crates.io page for feature flags and more information.

Overview

This crate contains:

Low level prost generated structs for Flight gRPC protobuf messages, such as FlightData, FlightInfo, Location and Ticket.

Low level tonic generated flight_service_client and flight_service_server.

Experimental support for Flight SQL in sql. Requires the flight-sql-experimental feature of this crate to be activated.

Re-exports

pub use client::FlightClient;

Modules

client

Mid Level FlightClient

decode

Decoder to create RecordBatch streams from FlightData streams. See FlightRecordBatchStream.

encode

Encoder to create FlightData streams from RecordBatch streams. See FlightDataEncoderBuilder.

error

Common error types

flight_descriptor

Defines a Flight for generation or retrieval.

flight_service_client

Low Level tonic FlightServiceClient.

flight_service_server

Low Level tonic FlightServiceServer and FlightService.

gen 🔒

sql

Support for execute SQL queries using Apache Arrow Flight SQL.

streams 🔒

FallibleRequestStream and FallibleTonicResponseStream adapters

trailers 🔒

Helper to extract HTTP/gRPC trailers from a tonic stream.

utils

Utilities to assist with reading and writing Arrow data as Flight messages

Structs

Action

An opaque action specific for the service.

ActionType

Describes an available action, including both the name used for execution along with a short description of the purpose of the action.

BasicAuth

A message for doing simple auth.

CancelFlightInfoRequest

The request of the CancelFlightInfo action.

CancelFlightInfoResult

The result of the CancelFlightInfo action.

Criteria

A service specific expression that can be used to return a limited set of available Arrow Flight streams.

Empty

FlightData

A batch of Arrow data as part of a stream of batches.

FlightDescriptor

The name or tag for a Flight. May be used as a way to retrieve or generate a flight or be used to expose a set of previously defined flights.

FlightEndpoint

A particular stream or split associated with a flight.

FlightInfo

The access coordinates for retrieval of a dataset. With a FlightInfo, a consumer is able to determine how to retrieve a dataset.

HandshakeRequest

The request that a client provides to a server on handshake.

HandshakeResponse

IpcMessage

IpcMessage represents a Schema in the format expected in FlightInfo.schema

Location

A location where a Flight service will accept retrieval of a particular stream given a ticket.

PollInfo

The information to process a long-running query.

PutResult

RenewFlightEndpointRequest

The request of the RenewFlightEndpoint action.

Result

An opaque result returned after executing an action.

SchemaAsIpc

SchemaAsIpc represents a pairing of a Schema with IpcWriteOptions

SchemaResult

Wrap the result of a getSchema call

Ticket

An opaque identifier that the service can use to retrieve a particular portion of a stream.

Enums

CancelStatus

The result of a cancel operation.

Functions

flight_schema_as_encoded_data 🔒

flight_schema_as_flatbuffer 🔒

limited_fmt 🔒

Limits the output of value to limit…

schema_to_ipc_format 🔒

Type Aliases

ArrowResult 🔒