arrow_flight
Module decode

Decoder to create RecordBatch streams from FlightData streams. See FlightRecordBatchStream.

Structs

DecodedFlightData

FlightData and the decoded payload (Schema, RecordBatch), if any

FlightDataDecoder

Wrapper around a stream of FlightData that handles the details of decoding low level Flight messages into [Schema] and [RecordBatch]es, including details such as dictionaries.

FlightRecordBatchStream

Decodes a [Stream] of FlightData back into [RecordBatch]es. This can be used to decode the response from an Arrow Flight server

FlightStreamState 🔒

tracks the state needed to reconstruct [RecordBatch]es from a streaming flight response.

Enums

DecodedPayload

The result of decoding FlightData