arrow_flight

Module encode

Encoder to create FlightData streams from RecordBatch streams. See FlightDataEncoderBuilder.

Structs

FlightDataEncoder

Stream that encodes a stream of record batches to flight data.

FlightDataEncoderBuilder

Creates a [Stream] of FlightDatas from a Stream of Result<RecordBatch, FlightError>.

FlightIpcEncoder 🔒

The data needed to encode a stream of flight data, holding on to shared Dictionaries.

Enums

DictionaryHandling

Defines how a FlightDataEncoder encodes DictionaryArrays

Constants

GRPC_TARGET_MAX_FLIGHT_SIZE_BYTES

Default target size for encoded FlightData.

Functions

hydrate_dictionaries 🔒

Hydrates any dictionaries arrays in batch to its underlying type. See hydrate_dictionary for more information.

hydrate_dictionary 🔒

Hydrates a dictionary to its underlying type.

prepare_field_for_flight 🔒

prepare_schema_for_flight 🔒

Prepare an arrow Schema for transport over the Arrow Flight protocol

split_batch_for_grpc_response 🔒

Split RecordBatch so it hopefully fits into a gRPC response.