arrow_flight

Module gen

Modules

flight_descriptor

Nested message and enum types in FlightDescriptor.

flight_service_client

Generated client implementations.

flight_service_server

Generated server implementations.

Structs

Action

An opaque action specific for the service.

ActionType

Describes an available action, including both the name used for execution along with a short description of the purpose of the action.

BasicAuth

A message for doing simple auth.

CancelFlightInfoRequest

The request of the CancelFlightInfo action.

CancelFlightInfoResult

The result of the CancelFlightInfo action.

Criteria

A service specific expression that can be used to return a limited set of available Arrow Flight streams.

Empty

FlightData

A batch of Arrow data as part of a stream of batches.

FlightDescriptor

The name or tag for a Flight. May be used as a way to retrieve or generate a flight or be used to expose a set of previously defined flights.

FlightEndpoint

A particular stream or split associated with a flight.

FlightInfo

The access coordinates for retrieval of a dataset. With a FlightInfo, a consumer is able to determine how to retrieve a dataset.

HandshakeRequest

The request that a client provides to a server on handshake.

HandshakeResponse

Location

A location where a Flight service will accept retrieval of a particular stream given a ticket.

PollInfo

The information to process a long-running query.

PutResult

RenewFlightEndpointRequest

The request of the RenewFlightEndpoint action.

Result

An opaque result returned after executing an action.

SchemaResult

Wrap the result of a getSchema call

Ticket

An opaque identifier that the service can use to retrieve a particular portion of a stream.

Enums

CancelStatus

The result of a cancel operation.