arrow_flight::sql::client

Struct FlightSqlServiceClient

pub struct FlightSqlServiceClient<T> {
    token: Option<String>,
    headers: HashMap<String, String>,
    flight_client: FlightServiceClient<T>,
}

A FlightSQLServiceClient is an endpoint for retrieving or storing Arrow data by FlightSQL protocol.

Fields
token: Option<String>headers: HashMap<String, String>flight_client: FlightServiceClient<T>
Implementations

impl FlightSqlServiceClient<Channel>

A FlightSql protocol client that can run queries against FlightSql servers This client is in the “experimental” stage. It is not guaranteed to follow the spec in all instances. Github issues are welcomed.

pub fn new(channel: Channel) -> Self

Creates a new FlightSql client that connects to a server over an arbitrary tonic Channel

pub fn new_from_inner(inner: FlightServiceClient<Channel>) -> Self

Creates a new higher level client with the provided lower level client

pub fn inner(&self) -> &FlightServiceClient<Channel>

Return a reference to the underlying FlightServiceClient

pub fn inner_mut(&mut self) -> &mut FlightServiceClient<Channel>

Return a mutable reference to the underlying FlightServiceClient

pub fn into_inner(self) -> FlightServiceClient<Channel>

Consume this client and return the underlying FlightServiceClient

pub fn set_token(&mut self, token: String)

Set auth token to the given value.

pub fn clear_token(&mut self)

Clear the auth token.

pub fn token(&self) -> Option<&String>

Share the bearer token with potentially different DoGet clients

pub fn set_header(&mut self, key: impl Into<String>, value: impl Into<String>)

Set header value.

async fn get_flight_info_for_command<M: ProstMessageExt>( &mut self, cmd: M, ) -> Result<FlightInfo, ArrowError>

pub async fn execute( &mut self, query: String, transaction_id: Option<Bytes>, ) -> Result<FlightInfo, ArrowError>

Execute a query on the server.

pub async fn handshake( &mut self, username: &str, password: &str, ) -> Result<Bytes, ArrowError>

Perform a handshake with the server, passing credentials and establishing a session.

If the server returns an “authorization” header, it is automatically parsed and set as a token for future requests. Any other data returned by the server in the handshake response is returned as a binary blob.

pub async fn execute_update( &mut self, query: String, transaction_id: Option<Bytes>, ) -> Result<i64, ArrowError>

Execute a update query on the server, and return the number of records affected

pub async fn execute_ingest<S>( &mut self, command: CommandStatementIngest, stream: S, ) -> Result<i64, ArrowError>
where S: Stream<Item = Result<RecordBatch>> + Send + 'static,

Execute a bulk ingest on the server and return the number of records added

pub async fn get_catalogs(&mut self) -> Result<FlightInfo, ArrowError>

Request a list of catalogs as tabular FlightInfo results

pub async fn get_db_schemas( &mut self, request: CommandGetDbSchemas, ) -> Result<FlightInfo, ArrowError>

Request a list of database schemas as tabular FlightInfo results

pub async fn do_get( &mut self, ticket: impl IntoRequest<Ticket>, ) -> Result<FlightRecordBatchStream, ArrowError>

Given a flight ticket, request to be sent the stream. Returns record batch stream reader

pub async fn do_put( &mut self, request: impl IntoStreamingRequest<Message = FlightData>, ) -> Result<Streaming<PutResult>, ArrowError>

Push a stream to the flight service associated with a particular flight stream.

pub async fn do_action( &mut self, request: impl IntoRequest<Action>, ) -> Result<Streaming<Result>, ArrowError>

DoAction allows a flight client to do a specific action against a flight service

pub async fn get_tables( &mut self, request: CommandGetTables, ) -> Result<FlightInfo, ArrowError>

Request a list of tables.

pub async fn get_primary_keys( &mut self, request: CommandGetPrimaryKeys, ) -> Result<FlightInfo, ArrowError>

Request the primary keys for a table.

pub async fn get_exported_keys( &mut self, request: CommandGetExportedKeys, ) -> Result<FlightInfo, ArrowError>

Retrieves a description about the foreign key columns that reference the primary key columns of the given table.

pub async fn get_imported_keys( &mut self, request: CommandGetImportedKeys, ) -> Result<FlightInfo, ArrowError>

Retrieves the foreign key columns for the given table.

pub async fn get_cross_reference( &mut self, request: CommandGetCrossReference, ) -> Result<FlightInfo, ArrowError>

Retrieves a description of the foreign key columns in the given foreign key table that reference the primary key or the columns representing a unique constraint of the parent table (could be the same or a different table).

pub async fn get_table_types(&mut self) -> Result<FlightInfo, ArrowError>

Request a list of table types.

pub async fn get_sql_info( &mut self, sql_infos: Vec<SqlInfo>, ) -> Result<FlightInfo, ArrowError>

Request a list of SQL information.

pub async fn get_xdbc_type_info( &mut self, request: CommandGetXdbcTypeInfo, ) -> Result<FlightInfo, ArrowError>

Request XDBC SQL information.

pub async fn prepare( &mut self, query: String, transaction_id: Option<Bytes>, ) -> Result<PreparedStatement<Channel>, ArrowError>

Create a prepared statement object.

pub async fn begin_transaction(&mut self) -> Result<Bytes, ArrowError>

Request to begin a transaction.

pub async fn end_transaction( &mut self, transaction_id: Bytes, action: EndTransaction, ) -> Result<(), ArrowError>

Request to commit/rollback a transaction.

pub async fn close(&mut self) -> Result<(), ArrowError>

Explicitly shut down and clean up the client.

fn set_request_headers<T>( &self, req: Request<T>, ) -> Result<Request<T>, ArrowError>