arrow_flight::sql::gen::command_statement_ingest

Module table_definition_options

Nested message and enum types in TableDefinitionOptions.

Enums

TableExistsOption

The action to take if the target table already exists

TableNotExistOption

The action to take if the target table does not exist