Module arrow_flight::sql::metadata::xdbc_info

Helpers for CommandGetXdbcTypeInfo metadata requests.

XdbcTypeInfo - a typed struct that holds the xdbc info corresponding to expected schema.

XdbcTypeInfoDataBuilder - a builder for collecting type infos and building a conformant RecordBatch.

XdbcTypeInfoData - a helper type wrapping a RecordBatch used for storing xdbc server metadata.

GetXdbcTypeInfoBuilder - a builder for consructing CommandGetXdbcTypeInfo responses.

Structs

GetXdbcTypeInfoBuilder

A builder for a CommandGetXdbcTypeInfo response.

XdbcTypeInfo

Data structure representing type information for xdbc types.

XdbcTypeInfoData

Helper to create CommandGetXdbcTypeInfo responses.

XdbcTypeInfoDataBuilder

Statics

GET_XDBC_INFO_SCHEMA 🔒

The schema for GetXdbcTypeInfo